Refactor FunctionTransformation for structs/traits.
